+++ src/edu/mtu/wup/ModelNoUi.java
@@ -7,6 +7,6 @@ public class ModelNoUi {
 	 * Main entry point for the model.
 	 */
 	public static void main(String[] args) {
-		ForestSim.load(HoughtonModel.class, args);
+		ForestSim.load(WupModel.class, args);
 	}
 }
