+++ src/edu/mtu/wup/ModelWithUI.java
@@ -5,7 +5,7 @@ import edu.mtu.simulation.ForestSimWithUI;
 
 public class ModelWithUI {
 	public static void main(String[] args) throws ForestSimException {
-		HoughtonModel model = new HoughtonModel(System.currentTimeMillis());
+		WupModel model = new WupModel(System.currentTimeMillis());
 		ForestSimWithUI fs = new ForestSimWithUI(model);
 		fs.load();
 	}
